+++ dflib-echarts/src/main/java/org/dflib/echarts/EChartHtmlSaver.java
@@ -1,6 +1,7 @@
 package org.dflib.echarts;
 
 import com.github.mustachejava.Mustache;
+import org.dflib.echarts.saver.HtmlChartModel;
 import org.dflib.echarts.saver.HtmlPageModel;
 
 import java.io.File;
@@ -8,19 +9,22 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.io.StringWriter;
 import java.io.Writer;
+import java.net.MalformedURLException;
+import java.net.URL;
 import java.nio.file.Path;
 import java.util.Arrays;
+import java.util.stream.Collectors;
 
 /**
  * @since 2.0.0
  */
 public class EChartHtmlSaver {
 
-    private static final Mustache HTML_PAGE_TEMPLATE = EChart.loadTemplate("html_page.mustache");
+    private static final Mustache DEFAULT_PAGE_TEMPLATE = EChart.loadTemplate("html_page.mustache");
 
     private boolean createMissingDirs;
     private String title;
-    private String chartStyle;
+    private URL htmlTemplate;
 
     /**
      * Instructs the saver to create any missing directories in the file path.
@@ -33,17 +37,61 @@ public class EChartHtmlSaver {
     }
 
     /**
-     * Overrides the default style of the chart "div" containers with the provided inline CSS snippet. The style argument
-     * will be inserted inside the <code>.dfl_ech { .. }</code> block and will be applied to each chart.
+     * Sets HTML page title.
      */
-    public EChartHtmlSaver chartStyle(String style) {
-        this.chartStyle = style;
-        return this;
-    }
-
     public EChartHtmlSaver title(String title) {
         this.title = title;
         return this;
+    }
+
+    /**
+     * Sets a custom HTML template. Can be filesystem based, a classpath resource or a web URL. If not set, a standard
+     * template is used. Custom template should contain the desired look-and-feel, with "dynamic" elements specified
+     * using the Mustache format. The saver will fill them with chart data:
+     * <ul>
+     *     <li>{{pageTitle}}</li>
+     *     <li>{{echartsUrl}}</li>
+     *     <li>{{chartDivN}}, where N is a number between 0 and 29</li>
+     *     <li>{{chartScriptN}}, where N is a number between 0 and 29</li>
+     *     <li>{{#charts}} / {{/charts}} - a list of charts</li>
+     *     <li>{{chartDiv}} - a chart div tag inside the {{#charts}} ... {{/charts}} block</li>
+     *     <li>{{chartScript}} - a chart script inside the {{#charts}} ... {{/charts}} block</li>
+     * </ul>
+     */
+    public EChartHtmlSaver htmlTemplate(URL htmlTemplate) {
+        this.htmlTemplate = htmlTemplate;
+        return this;
+    }
+
+    /**
+     * @see #htmlTemplate(URL)
+     */
+    public EChartHtmlSaver htmlTemplate(File htmlTemplate) {
+        try {
+            this.htmlTemplate = htmlTemplate.toURI().toURL();
+        } catch (MalformedURLException e) {
+            throw new RuntimeException(e);
+        }
+        return this;
+    }
+
+    /**
+     * @see #htmlTemplate(URL)
+     */
+    public EChartHtmlSaver htmlTemplate(Path htmlTemplate) {
+        try {
+            this.htmlTemplate = htmlTemplate.toUri().toURL();
+        } catch (MalformedURLException e) {
+            throw new RuntimeException(e);
+        }
+        return this;
+    }
+
+    /**
+     * @see #htmlTemplate(URL)
+     */
+    public EChartHtmlSaver htmlTemplate(String htmlTemplateFile) {
+        return htmlTemplate(new File(htmlTemplateFile));
     }
 
     public void save(Path path, EChartHtml... charts) {
@@ -79,21 +127,23 @@ public class EChartHtmlSaver {
 
     private void doSave(Appendable out, EChartHtml... charts) throws IOException {
 
-        String scriptImport = charts.length > 0 ? charts[0].getExternalScript() : "";
-
         HtmlPageModel model = new HtmlPageModel(
                 this.title != null ? this.title : "DFLib Chart",
-                scriptImport,
-                chartStyle != null ? chartStyle : "margin: 50px auto; padding: 20px;",
-                Arrays.asList(charts)
+                charts.length > 0 ? charts[0].getEchartsUrl() : "",
+                Arrays.stream(charts).map(c -> new HtmlChartModel(c.getChartDiv(), c.getChartScript())).collect(Collectors.toList())
         );
 
         boolean direct = out instanceof Writer;
         Writer writer = direct ? (Writer) out : new StringWriter();
-        HTML_PAGE_TEMPLATE.execute(writer, model);
+        htmlTemplate().execute(writer, model);
 
         if (!direct) {
             out.append(writer.toString());
         }
+    }
+
+    private Mustache htmlTemplate() {
+        // TODO: cache precompiled templates for cases of mass chart generation for a single template?
+        return this.htmlTemplate != null ? EChart.loadTemplate(htmlTemplate) : DEFAULT_PAGE_TEMPLATE;
     }
 }