+++ dflib-echarts/src/test/java/org/dflib/echarts/BarSeriesTest.java
@@ -11,20 +11,20 @@ public class BarSeriesTest {
     @Test
     public void test() {
 
-        String s2 = ECharts.chart().series(SeriesOpts.ofBar(), "y1").generateScriptHtml("_tid", df2);
+        String s2 = ECharts.chart().series(SeriesOpts.ofBar(), "y1").generateScript("_tid", df2);
         assertTrue(s2.contains("type: 'bar'"), s2);
 
-        String s3 = ECharts.chart().series(SeriesOpts.ofBar(), "y1").generateScriptHtml("_tid", df2);
+        String s3 = ECharts.chart().series(SeriesOpts.ofBar(), "y1").generateScript("_tid", df2);
         assertTrue(s3.contains("type: 'bar'"), s3);
     }
 
     @Test
     public void stack() {
 
-        String s2 = ECharts.chart().series(SeriesOpts.ofBar(), "y1").generateScriptHtml("_tid", df2);
+        String s2 = ECharts.chart().series(SeriesOpts.ofBar(), "y1").generateScript("_tid", df2);
         assertFalse(s2.contains("stack: "), s2);
 
-        String s3 = ECharts.chart().series(SeriesOpts.ofBar().stack(), "y1").generateScriptHtml("_tid", df2);
+        String s3 = ECharts.chart().series(SeriesOpts.ofBar().stack(), "y1").generateScript("_tid", df2);
         assertTrue(s3.contains("stack: 'total'"), s3);
     }
 
@@ -39,10 +39,10 @@ public class BarSeriesTest {
                 .borderType(LineType.dotted)
                 .opacity(0.55);
 
-        String s1 = ECharts.chart().series(SeriesOpts.ofBar(), "y1").generateScriptHtml("_tid", df2);
+        String s1 = ECharts.chart().series(SeriesOpts.ofBar(), "y1").generateScript("_tid", df2);
         assertFalse(s1.contains("itemStyle"), s1);
 
-        String s2 = ECharts.chart().series(SeriesOpts.ofBar().itemStyle(style), "y1").generateScriptHtml("_tid", df2);
+        String s2 = ECharts.chart().series(SeriesOpts.ofBar().itemStyle(style), "y1").generateScript("_tid", df2);
 
         assertTrue(s2.contains("type: 'bar'"), s2);
         assertTrue(s2.contains("itemStyle"), s2);