+++ dflib-echarts/src/test/java/org/dflib/echarts/EChartTest.java
@@ -11,14 +11,22 @@ public class EChartTest {
     public void plot() {
         EChartHtml ch = ECharts.chart().xAxis("x").series("y1", "y2").plot(df2);
         assertTrue(ch.getExternalScript().contains("<script type='text/javascript' src='https://cdn.jsdelivr.net/npm/echarts@5.5.1/dist/echarts.min.js'></script>"), ch.getExternalScript());
-        assertTrue(ch.getContainer().contains("<div id='dfl_ech_"), ch.getContainer());
-        assertTrue(ch.getScript().contains("['L0','A','B','C'],"), ch.getScript());
-        assertTrue(ch.getScript().contains("['y1',10,11,14],"), ch.getScript());
-        assertTrue(ch.getScript().contains("['y2',20,25,28]"), ch.getScript());
-        assertTrue(ch.getScript().contains("name: 'y1',"), ch.getScript());
-        assertTrue(ch.getScript().contains("name: 'y2',"), ch.getScript());
+        assertTrue(ch.getChartDiv().contains("<div id='dfl_ech_"), ch.getChartDiv());
+        assertTrue(ch.getChartScript().contains("['L0','A','B','C'],"), ch.getChartScript());
+        assertTrue(ch.getChartScript().contains("['y1',10,11,14],"), ch.getChartScript());
+        assertTrue(ch.getChartScript().contains("['y2',20,25,28]"), ch.getChartScript());
+        assertTrue(ch.getChartScript().contains("name: 'y1',"), ch.getChartScript());
+        assertTrue(ch.getChartScript().contains("name: 'y2',"), ch.getChartScript());
     }
 
+    @Test
+    public void scriptUrl() {
+
+        assertEquals("https://cdn.jsdelivr.net/npm/echarts@5.5.1/dist/echarts.min.js", ECharts.chart().scriptUrl());
+        assertEquals("https://example.org/x.js", ECharts.chart().scriptUrl("https://example.org/x.js").scriptUrl());
+    }
+
+    @Deprecated
     @Test
     public void generateExternalScriptHtml() {
 
@@ -33,40 +41,40 @@ public class EChartTest {
     public void generateContainerHtml() {
 
         String s1 = ECharts.chart().generateContainerHtml("_tid");
-        assertEquals("<div id='_tid' class='dfl_ech' style='width: 600px;height:400px;'></div>\n", s1);
+        assertEquals("<div id='_tid' class='dfl_ech' style='width: 600px;height:400px;'></div>", s1);
 
         String s2 = ECharts.chart().sizePx(20, 10).generateContainerHtml("_tid");
-        assertEquals("<div id='_tid' class='dfl_ech' style='width: 20px;height:10px;'></div>\n", s2);
+        assertEquals("<div id='_tid' class='dfl_ech' style='width: 20px;height:10px;'></div>", s2);
     }
 
     @Test
     public void darkTheme() {
 
-        String s1 = ECharts.chart().generateScriptHtml("_tid", df1);
+        String s1 = ECharts.chart().generateScript("_tid", df1);
         assertFalse(s1.contains("'dark'"), s1);
 
-        String s2 = ECharts.chart().darkTheme().generateScriptHtml("_tid", df1);
+        String s2 = ECharts.chart().darkTheme().generateScript("_tid", df1);
         assertTrue(s2.contains("'dark',"), s2);
     }
 
     @Test
     public void svgRenderer() {
 
-        String s1 = ECharts.chart().generateScriptHtml("_tid", df1);
+        String s1 = ECharts.chart().generateScript("_tid", df1);
         assertFalse(s1.contains("renderer: 'svg'"), s1);
 
-        String s2 = ECharts.chart().renderAsSvg().generateScriptHtml("_tid", df1);
+        String s2 = ECharts.chart().renderAsSvg().generateScript("_tid", df1);
         assertTrue(s2.contains("renderer: 'svg'"), s2);
     }
 
     @Test
     public void svgRenderer_DarkTheme() {
 
-        String s1 = ECharts.chart().generateScriptHtml("_tid", df1);
+        String s1 = ECharts.chart().generateScript("_tid", df1);
         assertFalse(s1.contains("'dark'"), s1);
         assertFalse(s1.contains("renderer: 'svg'"), s1);
 
-        String s2 = ECharts.chart().renderAsSvg().darkTheme().generateScriptHtml("_tid", df1);
+        String s2 = ECharts.chart().renderAsSvg().darkTheme().generateScript("_tid", df1);
         assertTrue(s2.contains("'dark',"), s2);
         assertTrue(s2.contains("renderer: 'svg'"), s2);
     }